+++ src/test/java/blackjack/domain/player/DealerTest.java
@@ -1,9 +1,9 @@
-package blackjack.domain;
+package blackjack.domain.player;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Letter;
 import blackjack.domain.card.Suit;
-import blackjack.domain.player.Player;
+import blackjack.domain.player.Dealer;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -13,14 +13,13 @@ import java.util.List;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-class PlayerTest {
+class DealerTest {
 
-    static Player player;
+    static Dealer dealer;
 
     @BeforeEach
     void setUp() {
-        player = new Player("IO", "3000",
-                new ArrayList<>(List.of(Card.of(Suit.DIAMOND, Letter.EIGHT), Card.of(Suit.HEART, Letter.JACK))));
+        dealer = new Dealer(new ArrayList<>(List.of(Card.of(Suit.DIAMOND, Letter.FIVE), Card.of(Suit.HEART, Letter.JACK))));
     }
 
     @Test
@@ -28,23 +27,23 @@ class PlayerTest {
     void receiveCardSuccess() {
         Card card = Card.of(Suit.CLOVER, Letter.FIVE);
 
-        player.hit(card);
+        dealer.hit(card);
 
-        List<Card> cards = player.getHand().getCards();
+        List<Card> cards = dealer.getHand().getCards();
         assertThat(cards.get(cards.size() - 1)).isEqualTo(card);
     }
 
     @Test
     @DisplayName("카드 새로 뽑을 수 있는지 판단 - 성공")
     void canReceiveNewCard() {
-        assertThat(player.isAbleToReceive()).isTrue();
+        assertThat(dealer.isAbleToReceive()).isTrue();
     }
 
     @Test
     @DisplayName("카드 새로 뽑을 수 있는지 판단 - 실패")
     void cannotReceiveNewCard() {
-        player.hit(Card.of(Suit.HEART, Letter.KING));
+        dealer.hit(Card.of(Suit.HEART, Letter.THREE));
 
-        assertThat(player.isAbleToReceive()).isFalse();
+        assertThat(dealer.isAbleToReceive()).isFalse();
     }
 }
