+++ src/test/java/blackjack/domain/player/NameTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.player;
 
 import blackjack.domain.player.Name;
 import org.assertj.core.api.Assertions;
