+++ src/test/java/blackjack/domain/player/PlayerTest.java
@@ -1,9 +1,9 @@
-package blackjack.domain;
+package blackjack.domain.player;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Letter;
 import blackjack.domain.card.Suit;
-import blackjack.domain.player.Dealer;
+import blackjack.domain.player.Player;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -13,13 +13,14 @@ import java.util.List;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-class DealerTest {
+class PlayerTest {
 
-    static Dealer dealer;
+    static Player player;
 
     @BeforeEach
     void setUp() {
-        dealer = new Dealer(new ArrayList<>(List.of(Card.of(Suit.DIAMOND, Letter.FIVE), Card.of(Suit.HEART, Letter.JACK))));
+        player = new Player("IO", "3000",
+                new ArrayList<>(List.of(Card.of(Suit.DIAMOND, Letter.EIGHT), Card.of(Suit.HEART, Letter.JACK))));
     }
 
     @Test
@@ -27,23 +28,23 @@ class DealerTest {
     void receiveCardSuccess() {
         Card card = Card.of(Suit.CLOVER, Letter.FIVE);
 
-        dealer.hit(card);
+        player.hit(card);
 
-        List<Card> cards = dealer.getHand().getCards();
+        List<Card> cards = player.getHand().getCards();
         assertThat(cards.get(cards.size() - 1)).isEqualTo(card);
     }
 
     @Test
     @DisplayName("카드 새로 뽑을 수 있는지 판단 - 성공")
     void canReceiveNewCard() {
-        assertThat(dealer.isAbleToReceive()).isTrue();
+        assertThat(player.isAbleToReceive()).isTrue();
     }
 
     @Test
     @DisplayName("카드 새로 뽑을 수 있는지 판단 - 실패")
     void cannotReceiveNewCard() {
-        dealer.hit(Card.of(Suit.HEART, Letter.THREE));
+        player.hit(Card.of(Suit.HEART, Letter.KING));
 
-        assertThat(dealer.isAbleToReceive()).isFalse();
+        assertThat(player.isAbleToReceive()).isFalse();
     }
 }
